% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{adjust}
\alias{adjust}
\title{Adjust to good dates}
\usage{
adjust(dates, bdc, calendar)
}
\arguments{
\item{dates}{a vector of dates to adjust.}

\item{bdc}{the business day convention used to roll the `dates` if necessary}

\item{calendar}{an object that inherits from [`Calendar`] or [`JointCalendar`]
which is used to determine the goodness of `dates`}
}
\value{
a vector of adjusted dates - good days are unadjusted
}
\description{
One common financial markets date arithmetic requires a date needs to be
rolled to the closest business day following some convention (see
[is_valid_bdc()] for further details). Such rolled dates can be determined by
calling `adjust()`.
}
\examples{
ausy <- AUSYCalendar()
adjust(lubridate::ymd("20120102"), "u", ausy)
adjust(lubridate::ymd("20120102"), "f", ausy)
adjust(lubridate::ymd("20120102"), "mf", ausy)
adjust(lubridate::ymd("20120102"), "p", ausy)
adjust(lubridate::ymd("20120102"), "mp", ausy)
adjust(lubridate::ymd("20120102"), "ms", ausy)
}
\seealso{
Other calendar methods: \code{\link{generate_schedule}},
  \code{\link{shift}}
}
