% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{annuity_fixed_leg}
\alias{annuity_fixed_leg}
\title{Title}
\usage{
annuity_fixed_leg(StartDate, EndDate, ValDate, cvDiscount, CouponFreq = "A",
  BusDayConv = "MF", DayCount = "30/360")
}
\arguments{
\item{BusDayConv}{}
}
\description{
Title
}
