% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{annuity_floating_leg}
\alias{annuity_floating_leg}
\title{Title}
\usage{
annuity_floating_leg(StartDate, EndDate, ValDate, cvDiscount, cvTenor,
  CouponFreq = "Q", BusDayConv = "MF", DayCount = "act/360")
}
\arguments{
\item{DayCount}{}
}
\description{
Title
}
