% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapCDS.R
\name{bootstrap_cds}
\alias{bootstrap_cds}
\title{Bootstrap a CDS curve for hazard rates}
\usage{

}
\arguments{
\item{MarketData}{CDS quotes for different Maturities}

\item{ZeroData}{A discount curve}

\item{ValuationDate}{Valuation date of the discount curve}

\item{Convention}{Convetion to use}

\item{OutputDates}{Date for output value}

\item{LGD}{Paramter for Loss-Given-Default}

\item{Output}{Type of Output}
}
\value{

}
\description{
Bootstrap a CDS curve for hazard rates
}
