% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapCDS.R
\name{BootstrapCDS}
\alias{BootstrapCDS}
\title{Title}
\usage{
BootstrapCDS(MarketData, ZeroData, ValuationDate, Convention,
  OutputDates = NULL, LGD, Output = "Hazard")
}
\arguments{
\item{Output}{}
}
\value{

}
\description{
Title
}
