% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{bump_curve}
\alias{bump_curve}
\title{Bump a given curve}
\usage{
bump_curve(Curve, Bump = 1e-04, IsDiscounts = TRUE)
}
\arguments{
\item{Curve}{A curve (discount or zero coupon)}

\item{Bump}{The bump to apply}

\item{IsDiscounts}{optional (default = TRUE) - indicates if the second column are discount facors or zero rates}
}
\value{
A bumped curve
}
\description{
Bump a given curve
}
\examples{
bump_curve(Curve)

}
