% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{Calendar}
\alias{Calendar}
\alias{EmptyCalendar}
\alias{AUSYCalendar}
\alias{AUMECalendar}
\alias{CHZHCalendar}
\alias{EUTACalendar}
\alias{GBLOCalendar}
\alias{HKHKCalendar}
\alias{JPTOCalendar}
\alias{NOOSCalendar}
\alias{NZAUCalendar}
\alias{NZWECalendar}
\alias{USNYCalendar}
\title{Build a calendar}
\usage{
Calendar(locale, tz)

EmptyCalendar()

AUSYCalendar()

AUMECalendar()

CHZHCalendar()

EUTACalendar()

GBLOCalendar()

HKHKCalendar()

JPTOCalendar()

NOOSCalendar()

NZAUCalendar()

NZWECalendar()

USNYCalendar()
}
\arguments{
\item{locale}{a four letter string representing an abbreviation of the
locale. The package uses locale representations loosely based on
[UN/LOCODE](http://www.unece.org/cefact/locode/welcome.html) (e.g.
Australia/Sydney is represented by `AUSY` rather than `AU/SYD` per the
LOCODE specification). The locale is used as a prefix to the calendar's
S3 class in the following manner: `<locale>Calendar` (e.g. `AUSYCalendar`).}

\item{tz}{the time zone associated with the given `locale` using
[OlsonNames()] (e.g. `Australia/Sydney`)}
}
\value{
`Calendar()` returns a function that constructs an object inheriting
  from the `Calendar` class. The calendar constructors provided by the
  package returns an object that inherits from `Calendar`.
}
\description{
Calendars are necessary for two reasons: they define whether a calendar day
is a good business day in a given locale and they are used to store the time
zone for the locale. Calendars can correspond to a single locale (usually a
city). These inherit from the `Calendar` class. The package implements a
number of calendars for key financial market locales such as
`AUSYCalendar`, `USNYCalendar` and `EUTACalendar` (TARGET). You can
also define a joint locale using [JointCalendar()].
}
\examples{
Calendar(NA, NA) # Defined: EmptyCalendar()
Calendar("AUSY", "Australia/Sydney") # Defined: AUSYCalendar()
}
\seealso{
Other calendar classes: \code{\link{JointCalendar}}
}
