% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{calibrate_hull_white_swaption}
\alias{calibrate_hull_white_swaption}
\title{Calibration of one-factor Hull-White parameters from Swaption volatilies}
\usage{
calibrate_hull_white_swaption(valdate, cvDiscount, cvTenor, vols)
}
\arguments{
\item{vols}{}
}
\description{
Calibration of one-factor Hull-White parameters from Swaption volatilies
}
