% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_frequency}
\alias{convert_frequency}
\title{Convert a numeric Tenor to a String}
\usage{
convert_frequency(input, atomic = FALSE)
}
\arguments{
\item{atomic}{}
}
\description{
Convert a numeric Tenor to a String
}
