% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{discount_to_rate}
\alias{discount_to_rate}
\title{Convert Discount Factors to Spot Rates}
\usage{
discount_to_rate(DF, T, Compounding = "CC")
}
\arguments{
\item{DF}{A discount factor (numeric)}

\item{Compounding}{optional (default="CC") - The compounding frequency}
}
\description{
Convert Discount Factors to Spot Rates
}
