% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{forward_rate}
\alias{forward_rate}
\title{Calculate forward rates}
\usage{
forward_rate(StartDate, EndDate, frame, DayCount = "act/360",
  compounding = "AC")
}
\arguments{
\item{compounding}{}
}
\description{
Calculate forward rates
}
