% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{forward_swap_rate}
\alias{forward_swap_rate}
\title{Title}
\usage{
forward_swap_rate(StartDate, EndDate, ValDate, cvDiscount, cvTenor,
  FloatingFreq = "q", FloatingDCC = "act/360", FloatingBDC = "mf",
  FixedFreq = "a", FixedDCC = "30/360", FixedBDC = "mf")
}
\arguments{
\item{FixedFreq}{}
}
\value{

}
\description{
Title
}
