% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{generate_dates}
\alias{generate_dates}
\title{Generating sequence of Dates}
\usage{

}
\arguments{
\item{StartDate}{Starting Date of the sequence}

\item{EndDate}{Ending Date of the sequence}

\item{ValDate}{optional (default=NULL) - Cuts of the date sequence.}

\item{SecondLast}{optional (default=NULL) - Adjust the second to last observation (See details).}

\item{CouponFreq}{optional (default="M") - Frequency of observations (See details).}

\item{BusDayConv}{optional (default="F") - Business Day Convention (See details).}

\item{IMM}{optional (default=FALSE) - Generates IMM dates (See details).}

\item{Output}{optional (default="Vector"). Type of output requested (See details).}
}
\value{
A vector or dateframe of Dates
}
\description{
Generating sequence of Dates
}
\details{
Generating sequency of dates moving backward from "End Date"
}
