% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDates.R
\name{GenerateDates}
\alias{GenerateDates}
\title{Title}
\usage{
GenerateDates(EffectiveDate, MaturityDate, ValDate = NULL,
  SecondLast = NULL, CouponFreq = "Monthly", BusDayConv = "F",
  IMM = FALSE, Output = "Vector")
}
\arguments{
\item{Output}{}
}
\description{
Title
}
