% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_white.R
\name{gurrieri_payer_swaption_closed_form}
\alias{gurrieri_payer_swaption_closed_form}
\title{Title}
\usage{
gurrieri_payer_swaption_closed_form(Strike, ValDate, TenorCurve, SwaptionDates,
  Tenor, Expiry, Notional = 1e+06, Alpha, S)
}
\arguments{
\item{S}{}
}
\description{
Title
}
