% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{interpolate}
\alias{interpolate}
\title{Interpolation}
\usage{
interpolate(X, Y, x, method = "linear")
}
\arguments{
\item{X, Y}{Vectors giving the coordinates of the points to be interpolated.
X can optionally be "Date" format,
allowing to interpolate on a typical j
discount curve structure.}

\item{x}{A single point specifying where to interpolate on observations of X.
Can be numeric, Date or character ("YYYY-MM-DD")}

\item{method}{Either linear ("l") or cubicspline ("cs")}
}
\value{
The value corresponding to x of values in Y.
}
\description{
Function useful for Interpolation.
   can be either "Dates" or "Numeric", allowing to
   easily Interpolate on discount curves.
   The interpolation can be done using linear or cubic-spline.
}
\examples{
Interpolate(X = c(1,3,5),Y = c(100,200,300),x = 2,method="l")
Interpolate(Curve$Date, Curve$Discount_Factors, "2025-03-03", "linear")

}
