% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{JointCalendar}
\alias{JointCalendar}
\title{Joint calendars}
\usage{
JointCalendar(calendars, rule = all)
}
\arguments{
\item{calendars}{a list of at least one `Calendar()` objects}

\item{rule}{either `all` or `any` corresponding to a date being good if
it is good in all or any of the calendars supplied.}
}
\value{
an object of class `JointCalendar` when using `JointCalendar()`
}
\description{
Sometimes the calendar governing a financial contract is defined by multiple
single locales. These joint calendars are represented by the `JointCalendar`
class.
}
\examples{
JointCalendar(list(AUSYCalendar(), AUMECalendar()), all)
JointCalendar(list(AUSYCalendar(), AUMECalendar()), any)
}
\seealso{
Other calendar classes: \code{\link{Calendar}}
}
