% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{parse_date}
\alias{parse_date}
\title{Parsing character to Dates class}
\usage{
parse_date(DateToParse, DateType = "European")
}
\arguments{
\item{DateToParse}{Input date value (integer/character/date)}

\item{DateType}{optional (default = "European") - format for DateToParse}
}
\value{
a Date
}
\description{
Parsing character to Dates class
}
