% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{parse_date_internal}
\alias{parse_date_internal}
\title{Parsing of dates to Excel numeric value}
\usage{
parse_date_internal(DateToParse, DateType = "European")
}
\arguments{
\item{DateToParse}{Input date value (integer/character/date)}

\item{DateType}{optional (default = "European") - format for DateToParse}
}
\value{
a Date in R date format
}
\description{
Parsing of dates to Excel numeric value
}
