% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{ParseDate}
\alias{ParseDate}
\title{Parsing character to Dates class}
\usage{
ParseDate(DateToParse, DateType = "European")
}
\arguments{
\item{DateToParse}{a character representing a date}

\item{DateType}{optional (default="European") - Format of DateToParse}
}
\value{
a Date
}
\description{
Parsing character to Dates class
}
