% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curves.R
\name{rate_to_discount}
\alias{rate_to_discount}
\title{Convert rates to discount factors}
\usage{
rate_to_discount(Rate, T, Compounding = "CC")
}
\arguments{
\item{Compounding}{}
}
\description{
Convert rates to discount factors
}
