% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{roll_month}
\alias{roll_month}
\title{Rolling months}
\usage{
roll_month(Date, Offset)
}
\arguments{
\item{Date}{a Date}

\item{Offset}{amount of months to roll}
}
\value{
a Date
}
\description{
Rolling months
}
