% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{roll_weekday}
\alias{roll_weekday}
\title{Rolls a date for a given Business Day Convention}
\usage{
roll_weekday(Day, BusDayConv = "F")
}
\arguments{
\item{Day}{A single Date or a vector of Dates}

\item{BusDayConv}{A specified business day convention}
}
\value{
A date in excel numeric format
}
\description{
Rolls a date for a given Business Day Convention
}
