% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{rollweekday}
\alias{rollweekday}
\title{Rolls a date for a given Business Day Convention}
\usage{
rollweekday(Day, BusDayConv = "F")
}
\arguments{
\item{Day}{A single Date or a vector of Dates}

\item{BusDayConv}{A specified business day convention}
}
\value{
A Date
}
\description{
Rolls a date for a given Business Day Convention
}
