% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuations.R
\name{value_fixed_leg}
\alias{value_fixed_leg}
\title{Title}
\usage{
value_fixed_leg(Curve, EffectiveDate, MaturityDate, SecondLast = NULL,
  Notional = 1e+07, ExchangeNotional = FALSE, Currency = "EUR",
  CouponRate = 0.01, Frequency = "A", DayCount = "ACT/360",
  DayConv = "MF", Direction = "PAY", Output = "DIRTY")
}
\arguments{
\item{Output}{}
}
\description{
Title
}
