% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuations.R
\name{value_floating_leg}
\alias{value_floating_leg}
\title{Title}
\usage{
value_floating_leg(CurveDiscount, CurveTenor, EffectiveDate, MaturityDate,
  SecondLast = NULL, LastFix = 0, Notional = 1e+07,
  ExchangeNotional = FALSE, Currency = "EUR", Tenor = "SA",
  Frequency = "SA", Spread = 0, DayConv = "MF", DayCount = "30/360",
  Direction = "RECEIVE", Output = "DIRTY")
}
\arguments{
\item{Output}{}
}
\description{
Title
}
