% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuations.R
\name{value_swaption}
\alias{value_swaption}
\title{Title}
\usage{
value_swaption(FwdRate, Strike, Vol, T, Annuity, Notional = 1e+05,
  Shift = 0, model = "BACH")
}
\arguments{
\item{model}{}
}
\description{
Title
}
