% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{yearfrac}
\alias{yearfrac}
\title{Calculate yearfraction between two Dates}
\usage{
yearfrac(DateBegin, DateEnd, DayCountConv = "ACT/360")
}
\arguments{
\item{DateBegin}{Starting Date}

\item{DateEnd}{Ending Date}

\item{DayCountConv}{Day Count Convention for yearfraction calculation
Default is ACT/360. See also 'Details'}
}
\value{
Numeric value representing years between Staring Date and Ending Date
}
\description{
Calculate yearfraction between two Dates
}
\details{
Dates can be inserted as character (DD-MM-YYYY or YYYY-MM-DD).

   Possible Day Count Convetions to use are "ACT/360","ACT/365, "ACT/ACT", "30/360".
   The argument is case-insensitive and "/" is optional. ("ACT360",...)
}
\examples{
yearfrac("2015-01-01","2020-01-01","act360")

}
