% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{bootstrap_cds}
\alias{bootstrap_cds}
\title{Bootstrap a CDS curve for hazard rates}
\usage{

}
\arguments{
\item{OutputDates}{Date for output value}

\item{LGD}{Paramter for Loss-Given-Default}
}
\value{

}
\description{
Bootstrap a CDS curve for hazard rates
}
