% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{date_to_excel}
\alias{date_to_excel}
\title{Convert a date in R to an excel numeric value}
\usage{
date_to_excel(d1)
}
\arguments{
\item{d1}{a Date in R}
}
\value{
a numeric value representing an Excel date
}
\description{
Every function exported to the user that returns a date
   must return a date value in Excel numeric value.
}
