% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{survival}
\alias{survival}
\title{Survival probability from Hazard rates}
\usage{
survival(SurvivalDate, ValuationDate, Maturities, Hazards,
  Convention = "act/360")
}
\arguments{
\item{SurvivalDate}{Date for surival chance}

\item{ValuationDate}{Date of calculation}

\item{Maturities}{Hazard rate maturities}

\item{Hazards}{Hazard rates}

\item{Convention}{Day count convention}
}
\value{
A survival percentage up to a given Date
}
\description{
Function takes a given date and Hazard rate Term Structure and calculates
}
