% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuations.R
\name{value_fixed_leg}
\alias{value_fixed_leg}
\title{Value a fixed leg}
\usage{
value_fixed_leg(Cv, ValD, EffD, MatD, SecLast = NULL, Not = 1e+07,
  ExchNot = FALSE, CCY = "EUR", Rate = 0.01, Freq = "A",
  DCC = "ACT/360", BDC = "MF", Dir = "PAY", Out = "DIRTY")
}
\arguments{
\item{Cv}{Discount Curve}

\item{ValD}{Valuation Date}

\item{EffD}{Effective Date}

\item{MatD}{Maturity Date}

\item{SecLast}{Second to last Payment Date}

\item{Not}{Notional}

\item{ExchNot}{Exchange of Notionals TRUE/FALSE}

\item{CCY}{Currency}

\item{Rate}{Coupon Rate}

\item{Freq}{Coupon Rate Frequency}

\item{DCC}{Day Count Convention}

\item{BDC}{Business Day Convention}

\item{Dir}{Direction PAY/SELL}

\item{Out}{Output type}
}
\value{
Depends on Output type
}
\description{
Value a fixed leg
}
