% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valuations.R
\name{value_floating_leg}
\alias{value_floating_leg}
\title{Title}
\usage{
value_floating_leg(CvDSC, CvTNR, ValD, EffD, MatD, SecLast = NULL,
  LstFix = 0, Not = 1e+07, ExchNot = FALSE, CCY = "EUR", Tenor = "SA",
  Freq = "SA", Spr = 0, BDC = "MF", DCC = "30/360", Dir = "RECEIVE",
  Out = "DIRTY")
}
\arguments{
\item{Tenor}{}
}
\description{
Title
}
