% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuations.R
\name{value_floating_leg}
\alias{value_floating_leg}
\title{Title}
\usage{
value_floating_leg(CvDSC, CvTNR, EffD, MatD, Stub = "short_front",
  LstFix = 0, Not = 1e+07, ExchNot = FALSE, Calendar = EUTACalendar(),
  CCY = "EUR", Reset = "T-2", Freq = "SA", Spr = 0, BDC = "MF",
  DCC = "30/360", Dir = "RECEIVE")
}
\arguments{
\item{Tenor}{}
}
\description{
Title
}
\examples{
value_floating_leg(CvDSC = EUR_OIS_20160930,CvTNR = EUR_6M_20160930,EffD = ymd(20120101),MatD = ymd(20210101))
}
